N059 autoClavicle ikFkSnap calculus
Cleaned code to be able to run the ikFkSnap with a good autoClavicle calculation, even if baking rotation to controllers.
